move XorAggregationResult implementation into own file
