+++ searchlib/src/vespa/searchlib/aggregation/xoraggregationresult.h
@@ -10,19 +10,24 @@ namespace search::aggregation {
 class XorAggregationResult : public AggregationResult {
 public:
     using Int64ResultNode = expression::Int64ResultNode;
+
+private:
+    Int64ResultNode _xor;
+
+public:
     DECLARE_AGGREGATIONRESULT(XorAggregationResult);
+
     void visitMembers(vespalib::ObjectVisitor& visitor) const override;
-    const Int64ResultNode& getXor() const { return _xor; }
+    [[nodiscard]] const Int64ResultNode& getXor() const { return _xor; }
     XorAggregationResult& setXor(const Int64ResultNode& i) {
         _xor = i;
         return *this;
     }
 
 private:
-    const ResultNode& onGetRank() const override { return getXor(); }
+    [[nodiscard]] const ResultNode& onGetRank() const override { return getXor(); }
     void onPrepare(const ResultNode& result) override;
     void initForUnitTest(const ResultNode& result) override;
-    Int64ResultNode _xor;
 };
 
 } // namespace search::aggregation
